Bug fixed.
If a comment spanned more than one line, removing it resulted in Java
code where the errors where not in the same lines as in the Java Decaf
code. Fixed by adding all the CRLF of the comment into the Java file,
so the lines are exactly in the same places (except for the HEADER, of
course).